:recycle: seeds: move certif platform configuration code
