+++ api/db/seeds/data/team-certification/data-builder.js
@@ -1,8 +1,9 @@
 import { ProSeed } from './cases/simple-pro-certification.js';
 import { ScoManagingStudent } from './cases/simple-sco-managing-students-certification.js';
-import { setupConfigurations } from './setup-configuration.js';
+import { setupConfigurations } from './shared/setup-configuration.js';
 
 async function teamCertificationDataBuilder({ databaseBuilder }) {
+  // Pix platform configuration
   await setupConfigurations({ databaseBuilder });
 
   // Cases
