+++ api/db/seeds/data/team-certification/tools/algorithm-configuration/create-flash-configuration.js
@@ -1,0 +1,10 @@
+export const createV3CertificationConfiguration = ({ databaseBuilder }) => {
+  databaseBuilder.factory.buildFlashAlgorithmConfiguration({
+    maximumAssessmentLength: 32,
+    challengesBetweenSameCompetence: null,
+    limitToOneQuestionPerTube: true,
+    enablePassageByAllCompetences: true,
+    variationPercent: 0.5,
+    createdAt: new Date('1977-10-19'),
+  });
+};
